@@ -1,14 +1,21 @@
 package com.nhom17.controllers;
 
+import com.nhom17.model.dto.ChiTietVe;
 import com.nhom17.model.dto.Customer;
+import com.nhom17.model.dto.GiaoDich;
 import com.nhom17.model.dto.PurchaseTicket;
+import com.nhom17.model.reposity.impl.GiaoDichDao;
 import com.nhom17.model.services.internal.database_interaction.DatabaseInteractionServiceFactory;
 import com.nhom17.model.services.internal.database_interaction.interfaces.BookingTicketService;
 
 import java.io.IOException;
+import java.sql.Time;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.time.LocalDate;
+import java.time.LocalTime;
 import java.util.Date;
+import java.util.List;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
@@ -28,62 +35,41 @@ public class PurchaseTicketControllerServlet extends BaseServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
 
-        String userName = request.getParameter("user-name");
-        String userEmail = request.getParameter("user-email");
-        String userPhone = request.getParameter("user-phone");
         HttpSession session = request.getSession(false);
         if (session != null) {
-            PurchaseTicket purchaseTicket = (PurchaseTicket) session.getAttribute("purchaseTicket");
+//            PurchaseTicket purchaseTicket = (PurchaseTicket) session.getAttribute("purchaseTicket");
+            GiaoDich giaoDich = (GiaoDich) session.getAttribute("giaoDich");
             String selectedSeats = (String) session.getAttribute("selectedSeats");
-            Customer customer = new Customer();
-            customer.setCustomerName(userName);
-            customer.setEmail(userEmail);
-            customer.setPhone(userPhone);
+            ChiTietVe chiTietVe = ChiTietVe.getByXuatChieu(giaoDich.getMaXuatChieu());
             System.out.println("----Purchase Ticket--------");
             System.out.println(
-                    "Hall No:" + purchaseTicket.getHallNo() + ",Ticket Price:" + purchaseTicket.getEachTicketPrice()
-                            + ",No. of Seats Reuest for booking:" + purchaseTicket.getNo_of_tickets());
+                    "Hall No:" + chiTietVe.getPhongChieu() + ",Ticket Price:" + giaoDich.getTongTien()
+                            + ",No. of Seats Reuest for booking:" + giaoDich.getListMaGhe().size());
             /*
              * System.out.println("Customer:" + customer.getCustomerName() +
              * ",Email:" + customer.getEmail() + "Phone:" +
              * customer.getPhone());
              */
-            System.out.println("ShowDate:" + purchaseTicket.getShowDate() + ",Time:" + purchaseTicket.getShowTime());
+            System.out.println("ShowDate:" + chiTietVe.getNgayChieu() + ",Time:" + chiTietVe.getGioChieu());
             System.out.println("Seat" + selectedSeats);
-            System.out.println("Movie Format:" + purchaseTicket.getMovieFormat() + "D");
-            System.out.println(purchaseTicket.getMovie());
-            System.out.println(customer.toString());
+            System.out.println("Movie Format:" + chiTietVe.getDangPhim());
             /*
              * int ticketId = dbConnector.bookSeats(customer.toString(),
              * selectedSeats, purchaseTicket.getShowDate(),
              * purchaseTicket.getShowTime(), purchaseTicket.getHallNo(),
              * purchaseTicket.getMovie(), purchaseTicket.getMovieFormat());
              */
-            String showDate = null;
-            try {
-                showDate = new SimpleDateFormat("dd-MMM-yyyy")
-                        .format(new SimpleDateFormat("dd-MMM-yy").parse(purchaseTicket.getShowDate()));
-            } catch (ParseException e) {
-                // TODO Auto-generated catch block
-                e.printStackTrace();
-            }
-            int ticketId = bookingTicketService.bookSeats(customer.toString(), selectedSeats, showDate,
-                    purchaseTicket.getShowTime(), purchaseTicket.getHallNo(), purchaseTicket.getMovie(),
-                    purchaseTicket.getMovieFormat());
-            /*
-             * int ticketId =
-             * dbConnector.bookSeats("Abc,abc@mail.com,01727412975",
-             * "U#F14,U#G12,U#G13,U#H13", "23-AUG-2016", "05: 00 PM", 2,
-             * "Independence Day: Resurgence", 2);
-             */
-            if (ticketId > 0) {
-                Date now = new Date();
-                purchaseTicket.setPurchaseTicketId(String.valueOf(ticketId));
-                purchaseTicket.setCustomer(customer);
-                purchaseTicket.setTotalCost(purchaseTicket.getEachTicketPrice() * purchaseTicket.getNo_of_tickets());
-                purchaseTicket.setPurchaseTime((new SimpleDateFormat("hh: mm a").format(now)).toUpperCase());
-                purchaseTicket.setPurchaseDate((new SimpleDateFormat("dd-MMM-yy").format(now)).toUpperCase());
+            int ticketId = 1;
+            GiaoDich giaoDich1 = GiaoDichDao.createGiaoDichDao().getByID(giaoDich.getId());
+            if (giaoDich1.getId() == giaoDich.getId()) {
+                giaoDich.setMaGiaoDich(giaoDich1.getMaGiaoDich());
+                giaoDich.setGioDat(LocalTime.now());
+                giaoDich.setNgayDat(LocalDate.now());
+                giaoDich.setIdTrangThaiGiaodich(2);
+                GiaoDichDao.createGiaoDichDao().update(giaoDich);
                 session.setAttribute("bookingStep", String.valueOf(4));
+                session.setAttribute("giaoDich", giaoDich);
+                session.setAttribute("chiTietVe", chiTietVe);
                 System.out.println("Ticket is sucessfully purchased");
                 String userTicket = request.getContextPath() + "/" + "final_ticket";
                 response.sendRedirect(userTicket);
